# docs/swagger.yaml
basePath: /
definitions:
...
  requests.CreateClassRequest:
    properties:
      hit_point_die_value:
        example: 4
        type: integer
      name:
        example: Barbarian
        maxLength: 200
        type: string
      primary_ability:
        example: Strength
        maxLength: 200
        type: string
      saves:
        example: '[''Strength'',''Constitution'']'
        maxLength: 200
        type: string
      short_description:
        example: Barbarians are mighty warriors who are powered by primal forces of
          the multiverse that manifest as a Rage.
        type: string
    required:
    - hit_point_die_value
    - name
    - primary_ability
    - saves
    - short_description
    type: object
...
  requests.UpdateClassRequest:
    properties:
      hit_point_die_value:
        example: 4
        type: integer
      name:
        example: Barbarian
        maxLength: 200
        type: string
      primary_ability:
        example: Strength
        maxLength: 200
        type: string
      saves:
        example: '[''Strength'',''Constitution'']'
        maxLength: 200
        type: string
      short_description:
        example: Barbarians are mighty warriors who are powered by primal forces of
          the multiverse that manifest as a Rage.
        type: string
    required:
    - hit_point_die_value
    - name
    - primary_ability
    - saves
    - short_description
    type: object
...
  responses.ClassPaginatedResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/responses.ClassResponse'
        type: array
      meta:
        $ref: '#/definitions/responses.ResponseMeta'
    type: object
  responses.ClassResponse:
    properties:
      hit_point_die_value:
        type: integer
      id:
        type: integer
      logo:
        $ref: '#/definitions/responses.FileResponse'
      name:
        type: string
      primary_ability:
        type: string
      saves:
        type: string
      short_description:
        type: string
      subclasses:
        items:
          $ref: '#/definitions/responses.SubclassResponse'
        type: array
    type: object
  responses.Data:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  responses.Error:
    properties:
      code:
        type: integer
      error:
        type: string
    type: object
  responses.FileResponse:
    properties:
      created_at:
        type: string
      enterprise_uid:
        type: string
      file_location:
        type: string
      filename:
        type: string
      id:
        example: 1
        type: integer
      model:
        type: string
      model_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  responses.ResponseMeta:
    properties:
      page:
        example: 3
        type: integer
      page_size:
        example: 25
        type: integer
      total_count:
        example: 300
        type: integer
    type: object
...
  responses.SubclassResponse:
    properties:
      class_id:
        type: integer
      id:
        type: integer
      logo:
        $ref: '#/definitions/responses.FileResponse'
      name:
        type: string
      short_description:
        type: string
    type: object
info:
  contact:
    email: jbrokenshire0306@gmail.com
    name: Jared Brokenshire
  description: A management solution for Dungeons & Dragons character sheet information
  title: D&D Character Dashboard
  version: 0.0.1
paths:
...
  /classes:
    get:
      consumes:
      - application/json
      description: List classes (paginated)
      operationId: classes-list
      parameters:
      - description: Search classes by name
        in: query
        name: search
        type: string
      - description: The page number
        in: query
        name: page
        type: integer
      - description: The numbers of items to return. Max 100
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ClassPaginatedResponse'
      summary: List classes
      tags:
      - Class Actions
    post:
      consumes:
      - application/json
      description: Create class
      operationId: classes-create
      parameters:
      - description: Class information
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.CreateClassRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.ClassResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Create class
      tags:
      - Class Actions
  /classes/{id}:
    delete:
      consumes:
      - application/json
      description: Delete class
      operationId: classes-delete
      parameters:
      - description: Class ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Data'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Delete class
      tags:
      - Class Actions
    get:
      consumes:
      - application/json
      description: Get class by ID
      operationId: classes-get
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ClassResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Get class by ID
      tags:
      - Class Actions
    put:
      consumes:
      - application/json
      description: Update class
      operationId: classes-update
      parameters:
      - description: Class ID
        in: path
        name: id
        required: true
        type: string
      - description: Class information
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateClassRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ClassResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Update class
      tags:
      - Class Actions
  /classes/{id}/upload/logo:
    post:
      consumes:
      - application/json
      description: Upload class logo
      operationId: classes-upload-logo
      parameters:
      - description: Class ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Data'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Upload class logo
      tags:
      - Class File Actions
...
swagger: "2.0"